EdkRepo: Centralize Reusable Unit Tests into a Shared Base Module
Changes:

- Created edkrepo_manifest_parser/test_bases/ package with base_tests.py
  containing all reusable test logic (~2825 lines).
- Removed edkrepo_manifest_parser/manifest_parser_unit_test_helpers/helpers.py,
  consolidating its content into the new centralized base_tests module.
- Updated all unit test files under edkrepo_manifest_parser/unit_tests/ to
  inherit from the corresponding base test classes, eliminating duplication.
- Deleted test_edk_manifest_validation.py, whose test cases have been migrated
  into the base_tests module.
- Updated TestCases.md documentation files across all tested components to
  reflect new test structure and coverage.
- Updated pytest.ini to include the new test_bases module in the test
  discovery configuration.

Fixes: #429
