+++ edkrepo_manifest_parser/__init__.py
@@ -3,7 +3,6 @@
 ## @file
 # __init__.py
 #
-# Copyright (c) 2017- 2019, Intel Corporation. All rights reserved.<BR>
+# Copyright (c) 2017 - 2026, Intel Corporation. All rights reserved.<BR>
 # SPDX-License-Identifier: BSD-2-Clause-Patent
 #
-
