+++ edkrepo_manifest_parser/unit_tests/FolderToFolderMappingFolder_TestCases.md
@@ -6,18 +6,22 @@
 Tests `_FolderToFolderMappingFolder.__init__` which parses optional `project1`/`project2` attributes and builds an excludes list from `<Exclude>` children.
 
 #### 1. All Fields Default When All Attributes and Children Absent
+- **Test Name**: `test_all_fields_default_when_absent`
 - **Description**: When all optional attributes (`project1`, `project2`) and `<Exclude>` child elements are absent.
 - **Expected Outcome**: `self.project1_folder` is `None`, `self.project2_folder` is `None`, and `self.excludes` is `[]`.
 
 #### 2. Sets project1_folder When project1 Attribute Present
+- **Test Name**: `test_init_sets_project_folder_when_present[project1_folder_present]`
 - **Description**: When the `project1` attribute is present.
 - **Expected Outcome**: `self.project1_folder` equals the attribute value.
 
 #### 3. Sets project2_folder When project2 Attribute Present
+- **Test Name**: `test_init_sets_project_folder_when_present[project2_folder_present]`
 - **Description**: When the `project2` attribute is present.
 - **Expected Outcome**: `self.project2_folder` equals the attribute value.
 
 #### 4. Populates excludes From Exclude Children
+- **Test Name**: `test_init_populates_excludes_from_exclude_children`
 - **Description**: When one `<Exclude>` child is present.
 - **Expected Outcome**: `self.excludes` contains one `_FolderToFolderMappingFolderExclude` with the correct path.
 
@@ -25,6 +29,7 @@ Tests `_FolderToFolderMappingFolder.__init__` which parses optional `project1`/`
 Tests `_FolderToFolderMappingFolder.tuple` which returns a `FolderToFolderMappingFolder` namedtuple.
 
 #### 1. Returns Correct FolderToFolderMappingFolder Namedtuple
+- **Test Name**: `test_tuple_returns_correct_folder_to_folder_mapping_folder_namedtuple`
 - **Description**: When both project folder attributes and one Exclude child are present.
 - **Expected Outcome**: `tuple` returns a `FolderToFolderMappingFolder` namedtuple with all fields correctly populated.
 
@@ -42,4 +47,3 @@ Tests `_FolderToFolderMappingFolder.tuple` which returns a `FolderToFolderMappin
    python3 -m pytest
    ```
    See the official `pytest` documentation at: https://docs.pytest.org/en/latest/how-to/usage.html for additional command line options.
-
